+++ engine/src/main/kotlin/engine/conditions/SignConditions.kt
@@ -1,0 +1,23 @@
+package engine.conditions
+
+import engine.expressions.Expression
+import engine.sign.Sign
+
+/**
+ * Returns true if the expression is definitely known to be non-positive (negative or zero),
+ * according to some heuristics
+ */
+fun Expression.isDefinitelyNotPositive(): Boolean = isDefinitely(Sign.NON_POSITIVE)
+
+fun Expression.isDefinitelyNotNegative(): Boolean = isDefinitely(Sign.NON_NEGATIVE)
+
+fun Expression.isDefinitelyNegative() = isDefinitely(Sign.NEGATIVE)
+
+fun Expression.isDefinitely(wantedSign: Sign): Boolean {
+    val sign = signOf()
+    return when {
+        sign.implies(wantedSign) -> true
+        sign.implies(wantedSign.negation()) -> false
+        else -> Sign.fromDouble(doubleValue).implies(wantedSign)
+    }
+}